Reorganize components directory structure
